@@ -1,16 +1,17 @@
 module github.com/datachainlab/cross
 
-go 1.15
+go 1.16
 
 require (
-	github.com/bluele/interchain-simple-packet v0.0.0-20200921011237-118864bc041e
+	github.com/bluele/interchain-simple-packet v0.0.0-20210621072258-be32aaced4b1
 	github.com/cosmos/cosmos-sdk v0.40.0-rc3
-	github.com/gogo/protobuf v1.3.1
+	github.com/gogo/protobuf v1.3.2
 	github.com/golang/protobuf v1.4.3
 	github.com/google/go-cmp v0.5.2 // indirect
 	github.com/gorilla/mux v1.8.0
 	github.com/grpc-ecosystem/grpc-gateway v1.15.2
 	github.com/rakyll/statik v0.1.7
+	github.com/regen-network/cosmos-proto v0.3.1 // indirect
 	github.com/spf13/cast v1.3.1
 	github.com/spf13/cobra v1.1.1
 	github.com/spf13/viper v1.7.1
