@@ -9,7 +9,7 @@ import (
 )
 
 func (p Keeper) ReceivePacket(ctx sdk.Context, packet channeltypes.Packet) (*sdk.Result, *packets.PacketAcknowledgementData, error) {
-	ip, err := packets.UnmarshalJSONIncomingPacket(p.m, packet)
+	ip, err := packets.UnmarshalIncomingPacket(p.m, packet)
 	if err != nil {
 		return nil, nil, sdkerrors.Wrapf(sdkerrors.ErrUnknownRequest, "unrecognized IBC packet type: %T: %v", packet, err)
 	}
@@ -24,8 +24,8 @@ func (p Keeper) ReceivePacket(ctx sdk.Context, packet channeltypes.Packet) (*sdk
 	return res, ack, nil
 }
 
-func (p Keeper) ReceivePacketAcknowledgementfunc(ctx sdk.Context, packet channeltypes.Packet, ack packets.IncomingPacketAcknowledgement) (*sdk.Result, error) {
-	ip, err := packets.UnmarshalJSONIncomingPacket(p.m, packet)
+func (p Keeper) ReceivePacketAcknowledgement(ctx sdk.Context, packet channeltypes.Packet, ack packets.IncomingPacketAcknowledgement) (*sdk.Result, error) {
+	ip, err := packets.UnmarshalIncomingPacket(p.m, packet)
 	if err != nil {
 		return nil, sdkerrors.Wrapf(sdkerrors.ErrUnknownRequest, "unrecognized IBC packet type: %T: %v", packet, err)
 	}
