@@ -1,7 +1,7 @@
 package packets
 
 import (
-	"encoding/json"
+	"encoding/hex"
 
 	sptypes "github.com/bluele/interchain-simple-packet/types"
 	"github.com/cosmos/cosmos-sdk/codec"
@@ -17,11 +17,15 @@ type (
 )
 
 // NewPacketData returns a new packet data
-func NewPacketData(h *Header, payloadJSON []byte) PacketData {
+func NewPacketData(h *Header, payload PacketDataPayload) PacketData {
 	if h == nil {
 		h = &Header{}
 	}
-	return sptypes.NewSimplePacketData(*h, payloadJSON)
+	any, err := utils.PackAny(payload)
+	if err != nil {
+		panic(err)
+	}
+	return sptypes.NewSimplePacketData(*h, any)
 }
 
 // PacketDataPayload defines the interface of packet data's payload
@@ -38,36 +42,28 @@ type PacketAcknowledgementPayload interface {
 	ValidateBasic() error
 }
 
-func MarshalJSONPacketData(data PacketData) ([]byte, error) {
-	bz, err := json.Marshal(data)
-	if err != nil {
-		return nil, err
-	}
-	return bz, nil
-}
-
-func UnmarshalJSONPacketData(bz []byte, pd *PacketData) error {
-	return json.Unmarshal(bz, pd)
-}
-
-func UnmarshalJSONIncomingPacket(m codec.Marshaler, raw exported.PacketI) (IncomingPacket, error) {
-	var pd PacketData
-	var payload PacketDataPayload
-	if err := UnmarshalJSONPacketData(raw.GetData(), &pd); err != nil {
-		return nil, err
-	}
-	if err := utils.UnmarshalJSONAny(m, &payload, pd.Payload); err != nil {
-		return nil, err
-	}
-	return NewIncomingPacket(raw, pd, payload), nil
-}
-
 // IncomingPacket defines the interface of incoming packet
 type IncomingPacket interface {
 	exported.PacketI
 	PacketData() PacketData
 	Header() Header
 	Payload() PacketDataPayload
+}
+
+func UnmarshalIncomingPacket(m codec.Marshaler, raw exported.PacketI) (IncomingPacket, error) {
+	data, err := hex.DecodeString(string(raw.GetData()))
+	if err != nil {
+		return nil, err
+	}
+	var pd PacketData
+	var payload PacketDataPayload
+	if err := proto.Unmarshal(data, &pd); err != nil {
+		return nil, err
+	}
+	if err := m.UnpackAny(pd.GetPayload(), &payload); err != nil {
+		return nil, err
+	}
+	return NewIncomingPacket(raw, pd, payload), nil
 }
 
 var _ IncomingPacket = (*incomingPacket)(nil)
@@ -105,7 +101,7 @@ func (p incomingPacket) Payload() PacketDataPayload {
 // OutgoingPacket defines the interface of outgoing packet
 type OutgoingPacket interface {
 	IncomingPacket
-	SetPacketData(m codec.JSONMarshaler, header Header, payload PacketDataPayload)
+	SetPacketData(header Header, payload PacketDataPayload)
 }
 
 var _ OutgoingPacket = (*outgoingPacket)(nil)
@@ -141,40 +137,32 @@ func (p outgoingPacket) Payload() PacketDataPayload {
 }
 
 // SetPacketData implements Outgoing.SetPacketData
-func (p *outgoingPacket) SetPacketData(m codec.JSONMarshaler, header Header, payload PacketDataPayload) {
+func (p *outgoingPacket) SetPacketData(header Header, payload PacketDataPayload) {
 	p.payload = payload
-	p.packetData = NewPacketData(&header, utils.MustMarshalJSONAny(m, payload))
+	p.packetData = NewPacketData(&header, payload)
 }
 
 // GetData implements Outgoing.GetData
 func (p outgoingPacket) GetData() []byte {
-	bz, err := MarshalJSONPacketData(p.packetData)
+	bz, err := proto.Marshal(&p.packetData)
 	if err != nil {
 		panic(err)
 	}
-	return bz
-}
-
-func MarshalJSONPacketAcknowledgementData(data PacketAcknowledgementData) ([]byte, error) {
-	bz, err := json.Marshal(data)
-	if err != nil {
-		return nil, err
-	}
-	return bz, nil
-}
-
-func UnmarshalJSONPacketAcknowledgementData(bz []byte, ad *PacketAcknowledgementData) error {
-	return json.Unmarshal(bz, ad)
+	return []byte(hex.EncodeToString(bz))
 }
 
 // NewPacketAcknowledgementData returns a new PacketAcknowledgementData
-func NewPacketAcknowledgementData(m codec.JSONMarshaler, h *Header, payload PacketAcknowledgementPayload) PacketAcknowledgementData {
+func NewPacketAcknowledgementData(h *Header, payload PacketAcknowledgementPayload) PacketAcknowledgementData {
 	if h == nil {
 		h = new(Header)
 	}
+	any, err := utils.PackAny(payload)
+	if err != nil {
+		panic(err)
+	}
 	return PacketAcknowledgementData{
 		Header:  *h,
-		Payload: utils.MustMarshalJSONAny(m, payload),
+		Payload: any,
 	}
 }
 
@@ -193,8 +181,8 @@ type incomingPacketAcknowledgement struct {
 var _ IncomingPacketAcknowledgement = (*incomingPacketAcknowledgement)(nil)
 
 // NewIncomingPacketAcknowledgement returns a new IncomingPacketAcknowledgement
-func NewIncomingPacketAcknowledgement(m codec.JSONMarshaler, h *Header, payload PacketAcknowledgementPayload) IncomingPacketAcknowledgement {
-	return incomingPacketAcknowledgement{data: NewPacketAcknowledgementData(m, h, payload), payload: payload}
+func NewIncomingPacketAcknowledgement(h *Header, payload PacketAcknowledgementPayload) IncomingPacketAcknowledgement {
+	return incomingPacketAcknowledgement{data: NewPacketAcknowledgementData(h, payload), payload: payload}
 }
 
 // Data implements IncomingPacketAcknowledgement.Data
@@ -212,24 +200,24 @@ func (a incomingPacketAcknowledgement) Payload() PacketAcknowledgementPayload {
 	return a.payload
 }
 
-func UnmarshalJSONIncomingPacketAcknowledgement(m codec.Marshaler, bz []byte) (IncomingPacketAcknowledgement, error) {
+func UnmarshalIncomingPacketAcknowledgement(m codec.Marshaler, ack []byte) (IncomingPacketAcknowledgement, error) {
 	var pd PacketAcknowledgementData
 	var payload PacketAcknowledgementPayload
 
-	if err := UnmarshalJSONPacketData(bz, &pd); err != nil {
+	if err := proto.Unmarshal(ack, &pd); err != nil {
 		return nil, err
 	}
-	if err := utils.UnmarshalJSONAny(m, &payload, pd.Payload); err != nil {
+	if err := m.UnpackAny(pd.GetPayload(), &payload); err != nil {
 		return nil, err
 	}
 
-	return NewIncomingPacketAcknowledgement(m, &pd.Header, payload), nil
+	return NewIncomingPacketAcknowledgement(&pd.Header, payload), nil
 }
 
 // OutgoingPacketAcknowledgement defines the interface of outgoing packet acknowledgement
 type OutgoingPacketAcknowledgement interface {
 	IncomingPacketAcknowledgement
-	SetData(m codec.JSONMarshaler, header Header, payload PacketAcknowledgementPayload)
+	SetData(header Header, payload PacketAcknowledgementPayload)
 }
 
 type outgoingPacketAcknowledgement struct {
@@ -238,9 +226,9 @@ type outgoingPacketAcknowledgement struct {
 }
 
 // NewOutgoingPacketAcknowledgement returns a new OutgoingPacketAcknowledgement
-func NewOutgoingPacketAcknowledgement(m codec.JSONMarshaler, h *Header, payload PacketAcknowledgementPayload) OutgoingPacketAcknowledgement {
+func NewOutgoingPacketAcknowledgement(h *Header, payload PacketAcknowledgementPayload) OutgoingPacketAcknowledgement {
 	return &outgoingPacketAcknowledgement{
-		data:    NewPacketAcknowledgementData(m, h, payload),
+		data:    NewPacketAcknowledgementData(h, payload),
 		payload: payload,
 	}
 }
@@ -263,7 +251,7 @@ func (a outgoingPacketAcknowledgement) Payload() PacketAcknowledgementPayload {
 }
 
 // Payload implements OutgoingPacketAcknowledgement.SetData
-func (a outgoingPacketAcknowledgement) SetData(m codec.JSONMarshaler, header Header, payload PacketAcknowledgementPayload) {
-	a.data = NewPacketAcknowledgementData(m, &header, payload)
+func (a outgoingPacketAcknowledgement) SetData(header Header, payload PacketAcknowledgementPayload) {
+	a.data = NewPacketAcknowledgementData(&header, payload)
 	a.payload = payload
 }
